# Conflicts:
#	gradle.properties
#	hapi-utils/src/main/java/com/hedera/services/ethereum/EthTxSigs.java
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/services/ServicesState.java
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/services/context/properties/BootstrapProperties.java
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/services/context/properties/GlobalDynamicProperties.java
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/services/context/properties/PropertySource.java
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/services/ledger/BalanceChange.java
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/services/ledger/accounts/staking/RewardCalculator.java
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/services/ledger/interceptors/AccountsCommitInterceptor.java
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/services/ledger/interceptors/StakingAccountsCommitInterceptor.java
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/services/state/migration/StateVersions.java
#	hedera-node/hedera-mono-service/src/main/resources/bootstrap.properties
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/services/ServicesStateTest.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/services/context/properties/BootstrapPropertiesTest.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/services/context/properties/GlobalDynamicPropertiesTest.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/services/context/properties/StandardizedPropertySourcesTest.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/services/ledger/accounts/staking/StakeChangeManagerTest.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/services/ledger/accounts/staking/StakeInfoManagerTest.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/services/ledger/interceptors/AccountsCommitInterceptorTest.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/services/ledger/interceptors/StakingAccountsCommitInterceptorTest.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/services/store/contracts/precompile/TransferPrecompilesTest.java
#	hedera-node/hedera-mono-service/src/test/resources/bootstrap.properties
#	hedera-node/hedera-mono-service/src/test/resources/bootstrap/standard.properties
#	settings.gradle.kts
#	test-clients/src/main/java/com/hedera/services/bdd/spec/queries/meta/HapiGetTxnRecord.java
#	test-clients/src/main/resource/bootstrap.properties

+++ test-clients/src/main/java/com/hedera/services/bdd/suites/contract/hapi/ContractCreateSuite.java
@@ -324,7 +324,7 @@ public class ContractCreateSuite extends HapiApiSuite {
 
     HapiApiSpec cannotSendToNonExistentAccount() {
         final var contract = "Multipurpose";
-        Object[] donationArgs = new Object[] {666666, "Hey, Ma!"};
+        Object[] donationArgs = new Object[] {666666L, "Hey, Ma!"};
 
         return defaultHapiSpec("CannotSendToNonExistentAccount")
                 .given(uploadInitCode(contract))
@@ -527,7 +527,12 @@ public class ContractCreateSuite extends HapiApiSuite {
                                             (int)
                                                     registry.getAccountID(bBeneficiary)
                                                             .getAccountNum();
-                                    final var sendArgs = new Object[] {sendAmount, aNum, bNum};
+                                    final var sendArgs =
+                                            new Object[] {
+                                                Long.valueOf(sendAmount),
+                                                Long.valueOf(aNum),
+                                                Long.valueOf(bNum)
+                                            };
 
                                     final var op =
                                             contractCall(contract, "sendTo", sendArgs)
@@ -654,9 +659,9 @@ public class ContractCreateSuite extends HapiApiSuite {
                                         contractCall(
                                                 sendInternalAndDelegateContract,
                                                 "sendRepeatedlyTo",
-                                                justSendContractNum.get(),
-                                                beneficiaryAccountNum.get(),
-                                                totalToSend / 2)),
+                                                BigInteger.valueOf(justSendContractNum.get()),
+                                                BigInteger.valueOf(beneficiaryAccountNum.get()),
+                                                BigInteger.valueOf(totalToSend / 2))),
                         getAccountBalance(beneficiary).hasTinyBars(totalToSend / 2),
                         /* But now we update the beneficiary to have a delegateContractId */
                         newKeyNamed(newKey)
@@ -669,9 +674,9 @@ public class ContractCreateSuite extends HapiApiSuite {
                                         contractCall(
                                                 sendInternalAndDelegateContract,
                                                 "sendRepeatedlyTo",
-                                                justSendContractNum.get(),
-                                                beneficiaryAccountNum.get(),
-                                                totalToSend / 2)),
+                                                BigInteger.valueOf(justSendContractNum.get()),
+                                                BigInteger.valueOf(beneficiaryAccountNum.get()),
+                                                BigInteger.valueOf(totalToSend / 2))),
                         getAccountBalance(beneficiary).hasTinyBars(3 * (totalToSend / 2)));
     }
 
@@ -706,9 +711,11 @@ public class ContractCreateSuite extends HapiApiSuite {
                                         contractCall(
                                                         sendInternalAndDelegateContract,
                                                         "sendRepeatedlyTo",
-                                                        justSendContractNum.get(),
-                                                        beneficiaryAccountNum.get(),
-                                                        balanceToDistribute / 2)
+                                                        BigInteger.valueOf(
+                                                                justSendContractNum.get()),
+                                                        BigInteger.valueOf(
+                                                                beneficiaryAccountNum.get()),
+                                                        BigInteger.valueOf(balanceToDistribute / 2))
                                                 .hasKnownStatus(INVALID_SIGNATURE)),
                         /* But it's not enough to just sign using an incomplete prefix */
                         sourcing(
@@ -716,9 +723,11 @@ public class ContractCreateSuite extends HapiApiSuite {
                                         contractCall(
                                                         sendInternalAndDelegateContract,
                                                         "sendRepeatedlyTo",
-                                                        justSendContractNum.get(),
-                                                        beneficiaryAccountNum.get(),
-                                                        balanceToDistribute / 2)
+                                                        BigInteger.valueOf(
+                                                                justSendContractNum.get()),
+                                                        BigInteger.valueOf(
+                                                                beneficiaryAccountNum.get()),
+                                                        BigInteger.valueOf(balanceToDistribute / 2))
                                                 .signedBy(DEFAULT_PAYER, beneficiary)
                                                 .hasKnownStatus(INVALID_SIGNATURE)),
                         /* We have to specify the full prefix so the sig can be verified async */
@@ -728,9 +737,11 @@ public class ContractCreateSuite extends HapiApiSuite {
                                         contractCall(
                                                         sendInternalAndDelegateContract,
                                                         "sendRepeatedlyTo",
-                                                        justSendContractNum.get(),
-                                                        beneficiaryAccountNum.get(),
-                                                        balanceToDistribute / 2)
+                                                        BigInteger.valueOf(
+                                                                justSendContractNum.get()),
+                                                        BigInteger.valueOf(
+                                                                beneficiaryAccountNum.get()),
+                                                        BigInteger.valueOf(balanceToDistribute / 2))
                                                 .alsoSigningWithFullPrefix(beneficiary)),
                         getAccountBalance(beneficiary).logged());
     }
